:bug: Manage DB container for RHSSO (#157)
These changes are intended to switch us from using a postgresql
container provisioned by the RHSSO operator which is unsupported and
uses a tag making it impossible to use MTA in a disconnected cluster
(without enabling an experimental feature to allow mirroring with tags).

The operator will cleanly switch the operator to use a postgresql
database container provisioned by us, but it will not migrate data.
Users will be required to log back with the default password and
reconfigure any users they have created.

~~Up to this point as far as I can tell RHSSO does not appear to
deprovision its postgresql deployment when the configuration is changed
to use an 'external' database and the keycloak-discovery service never
becomes ready, so for an upgrade these manual steps are currently
required. I'm still looking to see if we can smooth this out and have
some ideas. For example, we can probably look at the mta-rhsso keycloak
resource status for the keycloak-postgresql deployment and if it exists
nuke it and recreate it, which should clean up keycloak and reprovision
it properly.~~
  
~~oc delete keycloak -n openshift-mta mta-rhsso --ignore-not-found=true
&& \
oc scale deployment -n openshift-mta mta-operator --replicas=0 && \
oc scale deployment -n openshift-mta mta-operator --replicas=1~~

Signed-off-by: Jason Montleon <jmontleo@redhat.com>