:bug: Adjust Windup Addon Requests and Add Limits (#176)
Our requests are extremely inaccurate compared to what the container
can/will use. From what I've observed each container pegs a CPU at
pretty much 100%, and RAM usage far exceeds 2Gi.

The most immediate thing we can do to help prevent bulk analysis from
toppling nodes is to increase the requests for both of these.

Other things we should consider:
- Adding an env var that equals half the requests value the windup pod
to limit the JVM RAM usage. My belief that this should be half comes
from the description of MAX_MEMORY which reads as if additional NON-HEAP
Memory will be allocated equal to the HEAP memory specified by this var.

https://www.ibm.com/docs/en/coss/3.14.0?topic=application-change-memory-settings
```
Maximum memory the daemon should allocate for its heap when configuring memory dynamically.

Note: An equivalent amount will be used for off-heap memory unless configured otherwise....
```
This may effectively act as a limit so the next point may not be
necessary for constraining memory utilization.
- Implementing adjustable LIMITS on the windup pods in addition to
requests
- Linking to documentation and advising customers to implement eviction
thresholds to protect their nodes.

https://docs.openshift.com/container-platform/4.12/nodes/nodes/nodes-nodes-resources-configuring.html#allocate-eviction-thresholds_nodes-nodes-resources-configuring

I opened an issue on this last point because the eviction thresholds did
not work as it was suggested they would in the documentation.
https://issues.redhat.com/browse/OCPNODE-1579

Signed-off-by: Jason Montleon <jmontleo@redhat.com>