:bug: Use postgresql modules for migration (#364)
When using psql, errors in applying the sql generated from the dump do
not reach bash without adding additional options. psql only generates a
non-zero exit code when something fatal happens, like being unable to
connect to the database host. So, all ansible sees is exit code 0 from
psql and continues on, as if everything is great.

I did some reading up on postgres modules for ansible and we do not need
to shell out to do this, so I have implemented it with modules.

Since handling the actual migration with the ansible modules I have not
encountered a failure. Continuing to run upgrades to test ...

---------

Signed-off-by: Jason Montleon <jmontleo@redhat.com>